+++ src/Framework/FieldsAPI/Field.php
@@ -7,12 +7,14 @@ use Give\Framework\FieldsAPI\Contracts\Node;
 
 /**
  * @since 2.12.0
+ * @unreleased Support visibility conditions
  */
 abstract class Field implements Node {
 
 	use Concerns\HasDefaultValue;
 	use Concerns\HasName;
 	use Concerns\HasType;
+	use Concerns\HasVisibilityConditions;
 	use Concerns\IsReadOnly;
 	use Concerns\IsRequired;
 	use Concerns\SerializeAsJson;
@@ -21,6 +23,8 @@ abstract class Field implements Node {
 	protected $validationRules;
 
 	/**
+	 * @since 2.12.0
+	 *
 	 * @param string $name
 	 */
 	public function __construct( $name ) {